Big Changes
Added Man pages

fix Makezip
