+++ src/README.md
@@ -1,0 +1,28 @@
+# Todo
+
+- work on setHdmi to duplicate screen
+
+```sh
+xrandr --fb 1920x1080 --output DP-1 --mode 1680x1050 --scale-from 1920x1080
+--output DVI-D-1 --mode 1920x1080 --scale 1x1 --same-as DP-1
+```
+
+xrandr --output DVI-D-1 --same-as DP-1
+
+- [x] work on Engine::buildClangd in makehere
+
+- [x] work on a reimplementation of gitalias using libgit2
+
+- [x] work on gitalias-rust a rewrite of gitalias in rust
+
+- [] work on adding c++ tui lib to gitalias
+
+- [x] work on makehere , to create directory and add all the stuff required now cStrap
+
+- [x] add to gitalias default messages read from file in -m
+
+- [x] move fmt into gitalias
+
+- [x] remake msd
+
+- [x] work on docs for gitalias
