+++ src/makezip/includes.hpp
@@ -1,0 +1,20 @@
+#ifndef __MAKEZIP__INCLUDES_H_
+#define __MAKEZIP__INCLUDES_H_
+/* c++ libs */
+#include <dirent.h>
+#include <sys/stat.h>
+#include <zip.h>
+
+#include <ctime>
+/* boost for program options */
+#include <boost/program_options.hpp>
+#include <boost/program_options/errors.hpp>
+#include <boost/program_options/value_semantic.hpp>
+// #include <boost/property_tree/json_parser.hpp>
+#include <fmt/base.h>
+#include <fmt/format.h>
+#include <fmt/std.h>
+
+#include <boost/algorithm/string.hpp>
+#include <boost/program_options/parsers.hpp>
+#endif
