+++ src/setHdmi/Makefile
@@ -1,0 +1,14 @@
+FILE=_msd.cxx
+CC=g++
+fmt=-lfmt -std=c++2a -ggdb
+i=
+
+all: $(FILE)
+	$(CC) $(FILE) -o main $(fmt)
+
+run:
+	./main $(i)
+
+clean:
+	rm main
+
