--- src/underDevelopment/defines.h
@@ -1,10 +1,0 @@
-#ifndef DEFINES_H__
-#define DEFINES_H__
-
-// strings
-#define err "\e[31m[ ERR ]\e[0m"
-#define ok "\e[31m[ OK ]\e[0m"
-
-#define min(x, y) ((x) > (y) ? y : x)
-#define max(x, y) ((x) < (y) ? y : x)
-#endif
